% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corrPartialCounts}
\alias{corrPartialCounts}
\title{An innter loop of a generalized correlation sum. For internal use only.
No NAs assumed.}
\usage{
corrPartialCounts(x, r)
}
\description{
An innter loop of a generalized correlation sum. For internal use only.
No NAs assumed.
}

