% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corrSum}
\alias{corrSum}
\title{Returns a value of the correlation sum. No NAs assumed.}
\usage{
corrSum(x, r)
}
\arguments{
\item{x}{the vector}

\item{r}{distance coefficient}
}
\description{
Returns a value of the correlation sum. No NAs assumed.
}

