% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GrassProc.R
\name{corrSumG}
\alias{corrSumG}
\title{Returns a value of a generalized correlation sum with an additional q
parameter. Uses Rmpfr with precBits. Assumes no NAs in x.}
\usage{
corrSumG(x, r, q, precBits = 128)
}
\description{
Returns a value of a generalized correlation sum with an additional q
parameter. Uses Rmpfr with precBits. Assumes no NAs in x.
}

