% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GrassProc.R
\name{rCr}
\alias{rCr}
\title{Returns a data table representing a dependency between r and C(r)
correlation sum on k points over the [from, to] interval. Uses logs both
for r and C(r).}
\usage{
rCr(x, k = 100, from = -4, to = 0)
}
\description{
Returns a data table representing a dependency between r and C(r)
correlation sum on k points over the [from, to] interval. Uses logs both
for r and C(r).
}

