% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GrassProc.R
\name{rCrG}
\alias{rCrG}
\title{Generalized version of \code{rCr}. For variable values of q.
Uses Rmpfr with precBits. Assumes no NAs in x.}
\usage{
rCrG(x, k = 100, from = -4, to = 0, q, precBits = 128)
}
\description{
Generalized version of \code{rCr}. For variable values of q.
Uses Rmpfr with precBits. Assumes no NAs in x.
}

